Cloud Computing/Anabul Skin Disease Detection Backend/requirements.txt
Flask
gunicorn
tensorflow
numpy
pillow
matplotlib
keras
